backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
greenlet==3.0.1
psycopg2-binary==2.9.9
redis==5.0.1
asyncpg==0.29.0
aiosqlite==0.19.0

# Supabase
supabase==2.3.4

# Authentication and security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
slowapi==0.1.9
cryptography==41.0.7

# AI and ML
openai==1.3.7
google-generativeai==0.3.2
google-cloud-vision==3.4.4
Pillow==10.1.0

# Background tasks
celery==5.3.4

# Utilities
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
httpx==0.25.2
python-dateutil==2.8.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Web scraping
scrapy==2.11.0
selenium==4.15.2
webdriver-manager==4.0.1
playwright==1.40.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles
black==23.11.0
isort==5.12.0
flake8==6.1.0 